+++ CHANGELOG.md
@@ -53,6 +53,8 @@ to 2 and 3.
     - `PICODATA_DATA_DIR` renamed to `PICODATA_INSTANCE_DIR`
     - `PICODATA_ADMIN_SOCK` renamed to `PICODATA_ADMIN_SOCKET`
 
+- PgProto is now enabled by default and listens at `127.0.0.1:4327`.
+
 ### RPC API
 
 - `.proc_expel` and `.proc_expel_redirect` takes instance uuid instead of instance name.